pyproject.toml
[project]
name = "datainlet"
version = "0.0.0"
description = "A resident-centric open data platform to join datasets from different sources."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dagster>=1.10.4",
    "dagster-duckdb>=0.26.4",
    "dagster-duckdb-polars>=0.26.4",
    "duckdb>=1.2.1",
    "httpx>=0.28.1",
    "huggingface-hub>=0.29.3",
    "polars-lts-cpu>=1.29.0",
    "pyarrow>=19.0.1",
    "tenacity>=9.0.0",
    "dagster-webserver>=1.10.4",
    "ipykernel>=6.29.5",
    "ruff>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["datainlet"]

[tool.dagster]
module_name = "datainlet.definitions"
code_location_name = "datainlet"

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["win-64", "linux-64", "osx-64", "osx-arm64"]

[tool.pixi.tasks]
dev = { cmd = "dagster dev"}

[tool.pixi.dependencies]
python-pdal = ">=3.4.5,<4"
